Rename: classroom 라우팅 폴더 하위로 components폴더 이동
