+++ src/app/classroom/(components)/main/ContentCard.tsx
