+++ src/app/classroom/(components)/modal/common/LectureTitle.tsx
