+++ src/app/classroom/(components)/modal/createLecture/DropzoneSection.tsx
