+++ src/app/classroom/[lectureId]/(components)/lectureRoom/LectureComment.tsx
@@ -1,8 +1,8 @@
 import React, { FC, useState } from "react";
 import { useDispatch } from "react-redux";
 import { setModalVisibility } from "@/redux/slice/classroomModalSlice";
-import ReplySection from "@/components/classroomModal/comment/ReplySection";
-import CommentsSection from "@/components/lectureRoom/CommentsSection";
+import ReplySection from "@/app/classroom/(components)/modal/comment/ReplySection";
+import CommentsSection from "@/app/classroom/[lectureId]/(components)/lectureRoom/CommentsSection";
 import useGetComments from "@/hooks/lecture/useGetComments";
 
 interface LectureCommentProps {
