+++ src/app/classroom/[lectureId]/(components)/lectureRoom/LectureNavigation.tsx
