+++ src/app/classroom/[lectureId]/(components)/lectureRoom/typesOf/LinkLecture.tsx
