+++ src/app/classroom/[lectureId]/(components)/lectureRoom/typesOf/NoteLecture.tsx
