+++ src/app/classroom/[lectureId]/(components)/lectureRoom/typesOf/TypeOfLecture.tsx
