+++ src/app/classroom/[lectureId]/(components)/lectureRoom/typesOf/VideoLecture.tsx
